Feature/first data filter (#48)
* Add filter setup explanation

* Remove chapters at render

* Change timestamp to 1971-1976 for testing

* Remove some console.logs

* Remove fetch and add async loading of chapters

* Add chapters.js with async loading of data per chapter
