+++ views/new-story.ejs
@@ -19,9 +19,9 @@
 		<div id='map'></div>
 
 		<form class="location-and-timestamp" action="/create-story" method="post">
-			<input type="range" min="1900" max="2000" multiple value="1940,1980" />
-			<output id="rangevalueMin" name="timestamp-min">1940</output>
-			<output id="rangevalueMax" name="timestamp-max">1980</output>
+			<input type="range" min="1900" max="2000" multiple value="1971,1976" />
+			<output id="rangevalueMin" name="timestamp-min">1971</output>
+			<output id="rangevalueMax" name="timestamp-max">1976</output>
 
 			<button type="submit">Maak een verhaal</button>
 		</form>
